# 🚀 perf: reduce memory usage by using less stats object