@themgoncalves 2019-10-14
@endiliey thanks for your contribution! I will be running some tests and getting back with the results.

💪